refactor: restructure scan package — extract pipeline, consolidate files, simplify abstractions
Major architectural changes:
- Extract pipeline dispatch engine to scan/pipeline/ sub-package with Event interface
  (no generics, compatible with Go 1.17+)
- Merge engine/ + engines/ into unified scan/engine/ sub-package
- Move resources package from pkg/tools/resources/ to pkg/resources/

File consolidation (25 → 13 source files + 2 sub-packages):
- Merge finding.go + priority.go + event.go → event.go (unified domain model)
- Merge stats.go + projector.go + scan_data.go → collector.go
- Merge derive.go + adapters.go → adapter.go
- Merge profile.go → scan_options.go
- Merge verification_defaults.go → options.go
- Merge shared.go → input.go
- Rename scan_format.go → report.go, agent_verify.go → verify.go
- Split sdk_stage.go → engine/{gogo,spray,zombie,neutron}.go

Simplifications:
- Remove capabilitySpec abstraction layer — build capabilities inline
- Remove eventSink interface — use concrete collector
- Remove outputOptions struct — replace with bool parameter
- Remove serviceResult/webProbeResult/weakpassResult wrapper types
- Remove dead code (isSprayCapability) and unused pipeline constant aliases
- Decouple input parsing from collector (callback instead of *collector param)
- Add bridge.go for centralized scan↔pipeline type assertions
