+++ pkg/resources/embedded_stub.go
