+++ pkg/resources/pipeline_test.go
