+++ pkg/resources/template.go
