+++ pkg/tools/scan/scan_options.go
@@ -1,6 +1,9 @@
 package scan
 
-import "strings"
+import (
+	"fmt"
+	"strings"
+)
 
 const (
 	scanQuickDefaultPorts = "all"
@@ -83,4 +86,70 @@ func (o scanOptions) hasDiscoveryOverrides() bool {
 
 func (o scanOptions) hasWebOverrides() bool {
 	return len(o.Web.Dictionaries) > 0 || len(o.Web.Rules) > 0 || o.Web.Word != "" || o.Web.DefaultDict || o.Web.Advance
+}
+
+const (
+	scanModeQuick = "quick"
+	scanModeFull  = "full"
+)
+
+type profile struct {
+	Name          string
+	Capabilities  map[string]struct{}
+	CrawlDepth    int
+	AllowBroadPOC bool
+}
+
+func profileForMode(mode string) (profile, error) {
+	mode = strings.ToLower(strings.TrimSpace(mode))
+	if mode == "" {
+		mode = scanModeQuick
+	}
+
+	quickCaps := []string{
+		capGogoPortscan,
+		capSprayCheck,
+		capSprayFinger,
+		capCoreWeb,
+		capSprayCommon,
+		capSprayBackup,
+		capSprayActive,
+		capSprayCrawl,
+		capZombieWeakpass,
+		capNeutronPOC,
+	}
+
+	switch mode {
+	case scanModeQuick:
+		return profile{
+			Name:         scanModeQuick,
+			Capabilities: capabilitySet(quickCaps...),
+			CrawlDepth:   1,
+		}, nil
+	case scanModeFull:
+		fullCaps := append([]string{}, quickCaps...)
+		fullCaps = append(fullCaps,
+			capSprayBrute,
+		)
+		return profile{
+			Name:         scanModeFull,
+			Capabilities: capabilitySet(fullCaps...),
+			CrawlDepth:   2,
+		}, nil
+	default:
+		return profile{}, fmt.Errorf("unknown scan mode %q, expected quick or full", mode)
+	}
+}
+
+func capabilitySet(names ...string) map[string]struct{} {
+	out := make(map[string]struct{}, len(names))
+	for _, name := range names {
+		out[name] = struct{}{}
+	}
+	return out
+}
+
+func (p profile) Enabled(name string) bool {
+	_, ok := p.Capabilities[name]
+	return ok
 }
